requirements.txt
streamlit==1.44.1
pdfplumber==0.11.5
python-docx==1.1.2
